app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Luis Medina <info@lerman.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
django = "^4.2.5"
psycopg2-binary = "^2.9.9"
python-magic = "^0.4.27"
pandas = "^2.1.1"
openpyxl = "^3.1.2"
phonenumbers = "^8.13.22"
pytest = "^7.4.2"
black = "^23.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
